backend/requirements.txt
psycopg2-binary==2.8.6
gunicorn==20.0.4
idna==2.10
loguru==0.5.3
Django==3.2
djangorestframework==3.11.1
django-filter==21.1
requests==2.25.0